+++ src/app/routers/portfolio.py
@@ -1,13 +1,8 @@
 from fastapi import APIRouter, Query
 
 from app.contracts.portfolio import (
-    PortfolioAllocationResponse,
-    PortfolioBookResponse,
     PortfolioCatalogResponse,
     PortfolioInsightsResponse,
-    PortfolioLiquidityResponse,
-    PortfolioPositionBookResponse,
-    PortfolioProjectedCashflowResponse,
     PortfolioReadinessResponse,
     PortfolioWorkflowResponse,
     PortfolioWorkspaceResponse,
@@ -164,235 +159,3 @@ async def get_portfolio_workflow(
         correlation_id=correlation_id_var.get(),
         as_of_date=as_of_date,
     )
-
-
-@router.get(
-    "/portfolios/{portfolio_id}/book",
-    response_model=PortfolioBookResponse,
-    summary="Get portfolio book",
-    description=(
-        "Returns the combined portfolio book view used when the UI needs holdings, top "
-        "positions, allocation views, cash balances, and summary identity in one governed "
-        "response. Use this endpoint for a source-backed combined book snapshot instead of "
-        "reassembling those sections from separate contracts. The response keeps those book "
-        "sections aligned to one resolved as-of date so downstream clients do not need to "
-        "merge separate holdings, allocation, and liquidity reads."
-    ),
-)
-async def get_portfolio_book(
-    portfolio_id: str,
-    as_of_date: str | None = Query(
-        default=None,
-        description=(
-            "Optional as-of date in YYYY-MM-DD format used to resolve the combined book "
-            "snapshot across positions, allocations, and cash balances."
-        ),
-        examples=["2026-03-27"],
-    ),
-    include_projected: bool = Query(
-        default=False,
-        description=(
-            "Whether projected position rows should be included in the returned book when "
-            "upstream supports them."
-        ),
-        examples=[False],
-    ),
-    reporting_currency: str | None = Query(
-        default=None,
-        description=(
-            "Optional reporting currency used for book-level valuation restatement across "
-            "summary, holdings, allocations, and cash balances."
-        ),
-        examples=["USD"],
-    ),
-) -> PortfolioBookResponse:
-    return await portfolio_service().get_portfolio_book(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id_var.get(),
-        as_of_date=as_of_date,
-        include_projected=include_projected,
-        reporting_currency=reporting_currency,
-    )
-
-
-@router.get(
-    "/portfolios/{portfolio_id}/liquidity",
-    response_model=PortfolioLiquidityResponse,
-    summary="Get portfolio liquidity view",
-    description=(
-        "Returns the liquidity-focused portfolio view for cash balances, summary liquidity, "
-        "and projected cashflow. Use this endpoint when the UI needs current cash inventory "
-        "plus forward liquidity context without loading the full portfolio book. The "
-        "response preserves liquidity warnings and partial failures when forward cashflow "
-        "inputs are temporarily unavailable."
-    ),
-)
-async def get_portfolio_liquidity(
-    portfolio_id: str,
-    as_of_date: str | None = Query(
-        default=None,
-        description=(
-            "Optional as-of date in YYYY-MM-DD format used to resolve the liquidity snapshot "
-            "and forward cashflow context."
-        ),
-        examples=["2026-03-27"],
-    ),
-    reporting_currency: str | None = Query(
-        default=None,
-        description=(
-            "Optional reporting currency used for AUM and cash-balance restatement in the "
-            "liquidity view."
-        ),
-        examples=["USD"],
-    ),
-) -> PortfolioLiquidityResponse:
-    return await portfolio_service().get_portfolio_liquidity(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id_var.get(),
-        as_of_date=as_of_date,
-        reporting_currency=reporting_currency,
-    )
-
-
-@router.get(
-    "/portfolios/{portfolio_id}/projected-cashflow",
-    response_model=PortfolioProjectedCashflowResponse,
-    summary="Get portfolio projected cashflow view",
-    description=(
-        "Returns the forward-looking projected cashflow contract for a portfolio. "
-        "Use this endpoint when the UI needs a dedicated projected liquidity path for a "
-        "specific horizon without loading the broader liquidity summary. The response keeps "
-        "projection warnings and partial failures explicit when forward cashflow inputs are "
-        "temporarily degraded."
-    ),
-)
-async def get_portfolio_projected_cashflow(
-    portfolio_id: str,
-    as_of_date: str | None = Query(
-        default=None,
-        description=(
-            "Optional as-of date in YYYY-MM-DD format used to anchor the projected cashflow path."
-        ),
-        examples=["2026-03-27"],
-    ),
-    horizon_days: int = Query(
-        default=10,
-        ge=1,
-        le=365,
-        description="Forward projection horizon in business days for the requested cashflow path.",
-        examples=[30],
-    ),
-    include_projected: bool = Query(
-        default=True,
-        description=(
-            "Whether projected events should be included when deriving the forward cashflow path."
-        ),
-        examples=[True],
-    ),
-) -> PortfolioProjectedCashflowResponse:
-    return await portfolio_service().get_portfolio_projected_cashflow(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id_var.get(),
-        as_of_date=as_of_date,
-        horizon_days=horizon_days,
-        include_projected=include_projected,
-    )
-
-
-@router.get(
-    "/portfolios/{portfolio_id}/allocations",
-    response_model=PortfolioAllocationResponse,
-    summary="Get portfolio allocation views",
-    description=(
-        "Returns source-backed portfolio allocation views across the supported reporting "
-        "dimensions. Use this endpoint when the UI needs allocation buckets with optional "
-        "reporting-currency restatement and explicit look-through capability metadata. The "
-        "response preserves the effective look-through mode so downstream clients can tell "
-        "whether expanded exposure decomposition was actually applied."
-    ),
-)
-async def get_portfolio_allocations(
-    portfolio_id: str,
-    as_of_date: str | None = Query(
-        default=None,
-        description=(
-            "Optional as-of date in YYYY-MM-DD format used to resolve allocation views and "
-            "their summary framing."
-        ),
-        examples=["2026-03-27"],
-    ),
-    reporting_currency: str | None = Query(
-        default=None,
-        description=(
-            "Optional reporting currency used for allocation restatement across bucket "
-            "market values and weights."
-        ),
-        examples=["USD"],
-    ),
-    look_through_mode: str = Query(
-        default="direct_only",
-        description=(
-            "Requested allocation look-through mode for structured or fund exposures. "
-            "Use direct_only for booked exposures or full when downstream needs expanded "
-            "look-through buckets."
-        ),
-        examples=["direct_only", "full"],
-    ),
-) -> PortfolioAllocationResponse:
-    return await portfolio_service().get_portfolio_allocations(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id_var.get(),
-        as_of_date=as_of_date,
-        reporting_currency=reporting_currency,
-        look_through_mode=look_through_mode,
-    )
-
-
-@router.get(
-    "/portfolios/{portfolio_id}/positions",
-    response_model=PortfolioPositionBookResponse,
-    summary="Get portfolio positions view",
-    description=(
-        "Returns the detailed position book for a portfolio, including ranked top holdings. "
-        "Use this endpoint when the UI needs security-level holdings evidence, optional "
-        "projected rows, and reporting-currency-aware valuation fields. The response keeps "
-        "top holdings and full position rows aligned to one resolved position-book snapshot."
-    ),
-)
-async def get_portfolio_positions(
-    portfolio_id: str,
-    as_of_date: str | None = Query(
-        default=None,
-        description=(
-            "Optional as-of date in YYYY-MM-DD format used to resolve the position-book "
-            "snapshot and ranked top holdings."
-        ),
-        examples=["2026-03-27"],
-    ),
-    include_projected: bool = Query(
-        default=False,
-        description=(
-            "Whether projected position rows should be included when upstream supports them."
-        ),
-        examples=[False],
-    ),
-    reporting_currency: str | None = Query(
-        default=None,
-        description=(
-            "Optional reporting currency used for position valuation restatement across "
-            "market values and gain-loss fields."
-        ),
-        examples=["USD"],
-    ),
-) -> PortfolioPositionBookResponse:
-    return await portfolio_service().get_portfolio_positions(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id_var.get(),
-        as_of_date=as_of_date,
-        include_projected=include_projected,
-        reporting_currency=reporting_currency,
-    )
-
-
-
-
